+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `srl_summary_note` now gets added to all log records that pass through filter:
+  - Is only set whenever a message is printed and previous messages in the same stream were suppressed.
+
 ## [1.2.1] - 2022-12-07
 
 ### Fixed